Add insert update and delete functions

+++ src/main/java/university/app/dao/artistDAO.java
@@ -8,7 +8,7 @@ import java.util.Date;
 @Data
 @AllArgsConstructor
 public class artistDAO {
-    Integer id;
+    long id;
     String firstname;
     String secondname;
     String familyname;
